+++ docs/ReleaseNotes.md
@@ -4,6 +4,8 @@ _0.2.0_
 - Fix for exception on unknown format, now defaults to .NET default double format.
 - Fix for schema mappings falling back to ordinal not working.
 - Fix for empty shared string values created by AG Grid.
+- Fix handling of certain formats that would be misidentified as date-kind formats.
+- Fix Date formatting bug on netstandard2.0 where ticks values wouldn't be formatted as null characters.
 
 _0.1.12_
 - Fix for reading .xlsx files with inline-string values, believed to be created with Apache POI.